Added MainMenu, GUI Controller to MainController
Code now should be ported from GUI test to mainController to allow game
session to be created, joined, and to allow the game to be played.
Right now MainController begins in the main menu.  When the user presses
HostGame, he or she is currently taken to the game screen where the test
data and arena is drawn, but additional network setup may be required
before going to the game room.  Code needs to be implemented to have the
Maincontroller get the network playerData and then pass it on to the GUI
Controller to be drawn.  MainController can be run as the main to see
the menu improvements.